# backend/bracket/routes/players.py
from fastapi import APIRouter, Depends

from bracket.config import config
from bracket.database import DbConnection
from bracket.logic.ranking.players import get_player_statistics
from bracket.logic.ranking.statistics import PlayerStatistics
from bracket.logic.subscriptions import check_requirement
from bracket.models.db.player import Player, PlayerBody, PlayerMultiBody
from bracket.models.db.tournament import Tournament
from bracket.models.db.user import UserPublic
from bracket.routes.auth import (
    user_authenticated_for_tournament,
    user_authenticated_or_public_dashboard,
)
from bracket.routes.models import (
    PaginatedPlayers,
    PlayersResponse,
    SinglePlayerResponse,
    SuccessResponse,
)
from bracket.routes.util import disallow_archived_tournament
from bracket.schema import players
from bracket.sql.players import (
    get_all_players_in_tournament,
    get_player_count,
    insert_player,
    sql_delete_player,
)
from bracket.utils.db import fetch_one_parsed
from bracket.utils.id_types import PlayerId, TournamentId
from bracket.utils.pagination import PaginationPlayers
from bracket.utils.types import assert_some

router = APIRouter(prefix=config.api_prefix)


@router.get("/tournaments/{tournament_id}/players", response_model=PlayersResponse)
async def get_players(
    conn: DbConnection,
    tournament_id: TournamentId,
    not_in_team: bool = False,
    pagination: PaginationPlayers = Depends(),
    _: UserPublic | None = Depends(user_authenticated_or_public_dashboard),
) -> PlayersResponse:
    # The statistics follow from the match results, so they are worked out on every request
    # instead of being read from the players table.
    statistics = await get_player_statistics(conn, tournament_id)

    def with_statistics(player: Player) -> Player:
        return player.model_copy(update=statistics.get(player.id, PlayerStatistics()).model_dump())

    if pagination.sort_by in PlayerStatistics.model_fields:
        result = [
            with_statistics(player)
            for player in await get_all_players_in_tournament(
                conn, tournament_id, not_in_team=not_in_team
            )
        ]
        result.sort(
            key=lambda player: getattr(player, pagination.sort_by),
            reverse=pagination.sort_direction == "desc",
        )
        result = result[pagination.offset : pagination.offset + pagination.limit]
    else:
        result = [
            with_statistics(player)
            for player in await get_all_players_in_tournament(
                conn, tournament_id, not_in_team=not_in_team, pagination=pagination
            )
        ]

    return PlayersResponse(
        data=PaginatedPlayers(
            players=result,
            count=await get_player_count(conn, tournament_id, not_in_team=not_in_team),
        )
    )


@router.put("/tournaments/{tournament_id}/players/{player_id}", response_model=SinglePlayerResponse)
async def update_player_by_id(
    conn: DbConnection,
    tournament_id: TournamentId,
    player_id: PlayerId,
    player_body: PlayerBody,
    _: UserPublic = Depends(user_authenticated_for_tournament),
    __: Tournament = Depends(disallow_archived_tournament),
) -> SinglePlayerResponse:
    await conn.execute(
        players.update()
        .where((players.c.id == player_id) & (players.c.tournament_id == tournament_id))
        .values(**player_body.model_dump())
    )
    return SinglePlayerResponse(
        data=assert_some(
            await fetch_one_parsed(
                conn,
                Player,
                players.select().where(
                    (players.c.id == player_id) & (players.c.tournament_id == tournament_id)
                ),
            )
        )
    )


@router.delete("/tournaments/{tournament_id}/players/{player_id}", response_model=SuccessResponse)
async def delete_player(
    conn: DbConnection,
    tournament_id: TournamentId,
    player_id: PlayerId,
    _: UserPublic = Depends(user_authenticated_for_tournament),
    __: Tournament = Depends(disallow_archived_tournament),
) -> SuccessResponse:
    await sql_delete_player(conn, tournament_id, player_id)
    return SuccessResponse()


@router.post("/tournaments/{tournament_id}/players", response_model=SuccessResponse)
async def create_single_player(
    conn: DbConnection,
    player_body: PlayerBody,
    tournament_id: TournamentId,
    user: UserPublic = Depends(user_authenticated_for_tournament),
    _: Tournament = Depends(disallow_archived_tournament),
) -> SuccessResponse:
    existing_players = await get_all_players_in_tournament(conn, tournament_id)
    check_requirement(existing_players, user, "max_players")
    await insert_player(conn, player_body, tournament_id)
    return SuccessResponse()


@router.post("/tournaments/{tournament_id}/players_multi", response_model=SuccessResponse)
async def create_multiple_players(
    conn: DbConnection,
    player_body: PlayerMultiBody,
    tournament_id: TournamentId,
    user: UserPublic = Depends(user_authenticated_for_tournament),
    _: Tournament = Depends(disallow_archived_tournament),
) -> SuccessResponse:
    player_names = [player.strip() for player in player_body.names.split("\n") if len(player) > 0]
    existing_players = await get_all_players_in_tournament(conn, tournament_id)
    check_requirement(existing_players, user, "max_players", additions=len(player_names))

    for player_name in player_names:
        await insert_player(
            conn, PlayerBody(name=player_name, active=player_body.active), tournament_id
        )

    return SuccessResponse()