# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedPlayers": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/Player"
            },
            "title": "Players",
            "type": "array"
          }
        },
        "required": [
          "count",
          "players"
        ],
        "title": "PaginatedPlayers",
        "type": "object"
      },
...
      "Player": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "elo_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Elo Score",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "swiss_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Swiss Score",
            "type": "string"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "active",
          "name",
          "created",
          "tournament_id",
          "elo_score",
          "swiss_score",
          "wins",
          "draws",
          "losses",
          "id"
        ],
        "title": "Player",
        "type": "object"
      },
      "PlayerBody": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "name": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "active"
        ],
        "title": "PlayerBody",
        "type": "object"
      },
      "PlayerMultiBody": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "names": {
            "minLength": 1,
            "title": "Names",
            "type": "string"
          }
        },
        "required": [
          "names",
          "active"
        ],
        "title": "PlayerMultiBody",
        "type": "object"
      },
      "PlayersResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginatedPlayers"
          }
        },
        "required": [
          "data"
        ],
        "title": "PlayersResponse",
        "type": "object"
      },
...
      "SinglePlayerResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Player"
          }
        },
        "required": [
          "data"
        ],
        "title": "SinglePlayerResponse",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Bracket](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Rounds](#tag/Rounds)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Bracket, an open source tournament system.",
    "title": "Bracket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tournaments/{tournament_id}/players": {
      "get": {
        "operationId": "get_players_tournaments__tournament_id__players_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "not_in_team",
            "required": false,
            "schema": {
              "default": false,
              "title": "Not In Team",
              "type": "boolean"
            }
          },
          {
            "description": "Max number of results in a single page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Max number of results in a single page.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Filter results starting from this offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Filter results starting from this offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort_direction",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Sort Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "name",
              "enum": [
                "name",
                "elo_score",
                "swiss_score",
                "wins",
                "draws",
                "losses",
                "active",
                "created"
              ],
              "title": "Sort By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Players",
        "tags": [
          "Players"
        ]
      },
      "post": {
        "operationId": "create_single_player_tournaments__tournament_id__players_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Single Player",
        "tags": [
          "Players"
        ]
      }
    },
    "/tournaments/{tournament_id}/players/{player_id}": {
      "delete": {
        "operationId": "delete_player_tournaments__tournament_id__players__player_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Player",
        "tags": [
          "Players"
        ]
      },
      "put": {
        "operationId": "update_player_by_id_tournaments__tournament_id__players__player_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SinglePlayerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Player By Id",
        "tags": [
          "Players"
        ]
      }
    },
    "/tournaments/{tournament_id}/players_multi": {
      "post": {
        "operationId": "create_multiple_players_tournaments__tournament_id__players_multi_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerMultiBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Multiple Players",
        "tags": [
          "Players"
        ]
      }
    },
...
  }
}